Initial server
Functional REST ORM with compile-time guarantees. Current endpoints are
limited to /v1/user, with the appropriate endpoints being disabled in
production. Migrations allow for automated deploys with minimal issue,
and the build script will prevent a number of issues from being made. As
the existing endpoints have nothing specific to a user, I've extracted
it away into a macro, which can be used for additional endpoints (likely
GETting preset events at a minimum).

Using Rust instead of Node will provide a number of guarantees, and will
almost certainly be faster and more secure. Using Serde's attributes
allows me to entirely forget having to remove a user's refresh token,
and also enforces the types and presence (or lack of presence) of the
appropriate fields, something that was only partially done (and with
much effort) in Node.