Client wrapper for test, assert macros
This should make the testing more straightforward, as there's no longer
a `url` macro specific to each file. That has been replaced by the
`BASE` constant, which is then used by the new `Client` wrapper. The
instance methods don't require `.dispatch()`, and the parameters are
automatically cast into a `String` when necessary. This means we don't
have to call `.to_string()` on any JSON objects. Additionally, we can
directly use the values later on, without having to create temporary
variables.