Partial implementation of authenticated routes
`POST /v1/thread` now requires an authentication header to be passed,
which is then verified against the database. If the header does not
exist or is malformed, an appropriate HTTP response will be issued. Only
if the header is correct and verified will the route proceed.

Note that this is the _only_ endpoint that requires nothing more than an
authenticated user. Every other route requiring authentication requires
the user to be the thread author, local, or global admin. Of these,
global admin will be the easiest to implement, as it doesn't depend on
context. Because thread author and local admin _do_ depend on context,
the `User` object will have to be managed from inside the endpoint, not
solely as a request guard.

Testing gets a bit more verbose with authentication, as we're not
depending on any specific order. I've been careful to ensure the
appropriate teardowns are used so there's no extraneous rows remaining
after running tests. In the future, I may reorganize the tests to reduce
duplication and make things easier to follow.