Require a feature queryparam to be present to enable certain fields
This only affects what is _returned_ to the client. There are no
restrictions in place regarding updating of fields without the
queryparam set.

This should help Enceladus scale well, as there will no longer be an
excess of fields that aren't needed. Some existing fields will likely be
prefixed, ensuring they need to be explicitly requested as well.

Closes #3