Remove PresetEvent, GlobalAdmin
PresetEvent has been removed after consideration, as it essentially
served as a read-only store for the clients. As such, clients are now
expected to handle this on their own in whatever manner they see
appropriate.

GlobalAdmin was only being used on PresetEvent endpoints, so we can
safely remove it as well.