Allow an arbitrary number of columns on `Event`s
Currently, all columns must be PATCHed every time. There will be a
shorthand introduced in the future, likely by providing a JSON object
with the key being the index.

We also need to check that the values inserted are strictly strings or
numbers, as well as checking the length of the headers and fields are
the same. The latter may be possible using CHECKs in the database.