Move Reddit-related functionality to its own crate
This makes testing a bit odd, so I've added a mandatory parameter that
indicates if we're testing. This is necessary because `cfg!(test)`
doesn't appear to be shared between crates. There may be a better
solution, but that's for a later time. I presume Rust is able to
optimize this away with the release flag set.