Encrypt refresh, access tokens in database
Added a build script to automatically generate a key if one does not
exist. We only need to store the private, since the public can be
derived from that.

One thing I'd very much like to do is convert the new User structs to
a proc macro, though that could be difficult to generalize due to the
`#[cfg(test)]` attributes present. The implementations for `Into` will
need to be done by hand regardless.

Closes #2