Quasi-functional WebSocket server
Currently lets clients connect and join rooms, though no messages are
sent out. This is more a proof-of-concept, showing that it can be run
alongside Rocket in a separate thread.

A significant improvement on the structuring of WebSocket rooms is
currently pending on the upstream housleyjk/ws-rs#257 to be merged. That
will allow the use of a `HashSet` instead of using a `LinkedList`,
providing O(1) deletions rather than O(n).

Another upstream that is pending (though not directly affecting the
build) is rust-lang/rust-clippy#3844. This has been merged and is
pending release. Once this is released, the numerous warnings regarding
`const fn` from clippy should be silenced, as it's either external or
not currently possible in Rust.